drivers/gpio/gpio-nomadik.c: fix number of gpio per chip
In nmk_gpio_probe(), 'chip->ngpio' was given the value of 'num_gpio'
member of platform data, which happens to be 0 because it is not
initialized.
To overcome this, I set up the correct value (NMK_GPIO_PER_CHIP) in
the 'nmk_gpio_template' data structure, and I deleted the assigment
from platform data.